Move handle typing tests into a new class (and a new file)
